first draft of updating url patterns
