+++ src/pages/research/collections/online-resources/search.js
@@ -1,17 +1,19 @@
 import React, { Fragment } from 'react';
 // Apollo
 import { getDataFromTree } from '@apollo/client/react/ssr';
-import { withApollo } from './../../../apollo/client/withApollo';
+import { withApollo } from '../../../../apollo/client/withApollo';
 // Redux
-import { withRedux } from './../../../redux/withRedux';
+import { withRedux } from '../../../../redux/withRedux';
 // Components
-import PageContainer from './../../../components/shared/layouts/PageContainer';
-import RightRail from './../../../components/shared/RightRail';
-import SearchHeader from './../../../components/shared/SearchHeader';
-import SearchForm from './../../../components/online-resources/SearchForm';
-import SearchResults from './../../../components/online-resources/SearchResults';
-import SidebarMenus from './../../../components/online-resources/SidebarMenus';
-import Hero from './../../../components/online-resources/Hero';
+import PageContainer from '../../../../components/shared/layouts/PageContainer';
+import RightRail from '../../../../components/shared/RightRail';
+import SearchHeader from '../../../../components/shared/SearchHeader';
+import SearchForm from '../../../../components/online-resources/SearchForm';
+import SearchResults from '../../../../components/online-resources/SearchResults';
+import SidebarMenus from '../../../../components/online-resources/SidebarMenus';
+import Hero from '../../../../components/online-resources/Hero';
+// Utils
+import { ONLINE_RESOURCES_BASE_PATH } from './../../../../utils/config';
 
 function OnlineResourcesSearchPage() {
   return (
@@ -19,7 +21,7 @@ function OnlineResourcesSearchPage() {
       metaTags={{
         title: 'Online Resources',
         description: 'Hello welcome to the NYPL!',
-        url: 'https://www.nypl.org/research/online-resources'
+        url: `https://www.nypl.org${ONLINE_RESOURCES_BASE_PATH}/search`
       }}
       wrapperClass='nypl--research'
       contentHeader={
